@@ -2,8 +2,8 @@ package stark.stellasearch.dto.params;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 
 @Data
 public class CreatePlaylistRequest
