@@ -2,8 +2,8 @@ package stark.stellasearch.dto.params;
 
 import lombok.Data;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
 public class NewVideoUploadingTaskRequest
