@@ -10,9 +10,9 @@ import stark.dataworks.basic.exceptions.ExceptionInfoFormatter;
 import stark.dataworks.boot.web.ServiceResponse;
 import stark.stellasearch.dto.results.LoginStateToken;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
