@@ -16,10 +16,10 @@ import stark.stellasearch.service.dto.User;
 import stark.stellasearch.service.redis.RedisKeyManager;
 import stark.stellasearch.service.redis.StellaRedisOperation;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
